handle the re-authorization and re-subscription from the init state machine.  Handle this differently through the user of re-useable workers.
Also includes changes to gloox in order to get the subscription list.
gloox did not handle this correctly.  It was keeping a map of
a subscription node and its corresponding subscription.  It did not
support a subscription node and multiple subscriptions, like
we retrieve from the server.  Changed gloox code to have list
for a node.   Passed on changes to Jakob gloox author for him
to review.